+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Deprecation log configuration support added.
 
 #### Bugfixes
+* Fixed case in which index template could prepend an additional 'index.' to index settings.
 
 #### Changes
 * sysctl settings are no longer managed by the thias/sysctl module.